fix: create .env iff it does not already exist

+++ .devcontainer/post-create.sh
@@ -3,4 +3,6 @@ pnpm i
 
 # Configure environment variables
 # There is an `.env.example` in the root directory you can use for reference
-cp .env.example .env
+if [ ! -f .env ]; then
+  cp .env.example .env
+fi
